Bug Fix: Jump Forward.
In the original game, pressing Jump + (Left || Right) should allow Lara
to jump forward too. In OpenLara this is not possible as the input code
only checks if forward flag is set when pressed. Not forward + (Left ||
Right) resuling in Lara not being able to jump forward and only straight
up in these cases.