Update it.h (#255)
Sorry my friend, I translated again some lines comparing my work with the french FMVs of the game, in order to understand the meaning of certain slangs; unluckily you were faster than me. :) 
This translation is the best. Sorry for editing too much time the file, but I'm a perfectionist! :)
This is definitely the last revision.