restore high graphics detail on macOS
The macOS platform currently creates and uses an OpenGL 2.1 profile,
but it can still achieve high graphics detail in theory. In practice,
that ability regressed.

Prior to PR #355, extSupport made use of an undeclared glGetStringi,
a build error. That was fixed by importing a header that declared it.
That caused extSupport to start reporting false for any input. This
is due to the (glGetStringi != NULL) check not having the intended
effect on macOS, as reported in #357.

As long as OpenGL 2.1 profile is used, extSupport can't make use
of the glGetStringi + GL_NUM_EXTENSIONS code path, so explicitly
disable it on macOS for now via a preprocessor guard.

Another regression introduced by #355 is the removal of redefines
of glGenVertexArrays, glDeleteVertexArrays, glBindVertexArray with
Apple-specific versions, needed to get VAO alongside OpenGL 2.1.
So, bring this back, along with the comment justifying why they're
needed (as was done before 6e6ed3b2d0e8c0c9af83d5cb9653793ca76047f4).
Import OpenGL/glext.h to get definitions of those APPLE functions.

With these changes, high graphics detail settings are available on
macOS once again.

Fixes #356.
Updates #357.